Huge refactor to math system, with Vector3D becoming a template class, to support double type components. Implemented a font importer. Other misc changes

+++ include/mathernogl/maths/RaycastOperations.h
@@ -11,8 +11,8 @@ namespace mathernogl {
 struct Raycast{
 	Vector3D position;
 	Vector3D direction;
-	float minTValue = -std::numeric_limits<float>::infinity();
-	float maxTValue = std::numeric_limits<float>::infinity();
+	float minTValue = -std::numeric_limits<double>::infinity();
+	float maxTValue = std::numeric_limits<double>::infinity();
 };
 
 class mgl_no_intersection : public std::runtime_error {
@@ -20,6 +20,6 @@ public:
 	mgl_no_intersection(std::string message) : std::runtime_error(message) {}
 };
 
-Vector3D getRaycastIntersectionOnXZPlane(const Raycast& raycast, const float planeOffset);
+Vector3D getRaycastIntersectionOnXZPlane(const Raycast& raycast, const double planeOffset);
 
 }
